gadget: preserve files when indicated by content change observer
The observer can indicate that a file should be preserved during update/write,
even if it is not listed in the `preserve` files list inside gadget.yaml file.
Adjust the backup/update/rollback passes so that the requests are respected.

+++ gadget/mountedfilesystem.go
@@ -31,7 +31,6 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/snapcore/snapd/bootloader"
 	"github.com/snapcore/snapd/logger"
 	"github.com/snapcore/snapd/osutil"
 	"github.com/snapcore/snapd/strutil"
@@ -57,21 +56,20 @@ func checkContent(content *VolumeContent) error {
 	return nil
 }
 
-func observe(observer ContentObserver, op ContentOperation, ps *LaidOutStructure, root, dst string, data *ContentChange) error {
+func observe(observer ContentObserver, op ContentOperation, ps *LaidOutStructure, root, dst string, data *ContentChange) (ContentChangeAction, error) {
 	if observer == nil {
-		return nil
+		return ChangeApply, nil
 	}
 	relativeTarget := dst
 	if strings.HasPrefix(dst, root) {
 		// target path isn't really relative, make it so now
 		relative, err := filepath.Rel(root, dst)
 		if err != nil {
-			return err
+			return ChangeAbort, err
 		}
 		relativeTarget = relative
 	}
-	_, err := observer.Observe(op, ps, root, relativeTarget, data)
-	return err
+	return observer.Observe(op, ps, root, relativeTarget, data)
 }
 
 // TODO: MountedFilesystemWriter should not be exported
@@ -197,8 +195,12 @@ func (m *MountedFilesystemWriter) observedWriteFileOrSymlink(volumeRoot, src, ds
 		// with content in this file
 		After: src,
 	}
-	if err := observe(m.observer, ContentWrite, m.ps, volumeRoot, dst, data); err != nil {
+	act, err := observe(m.observer, ContentWrite, m.ps, volumeRoot, dst, data)
+	if err != nil {
 		return fmt.Errorf("cannot observe file write: %v", err)
+	}
+	if act == ChangePreserveBefore && osutil.FileExists(dst) {
+		return nil
 	}
 	return writeFileOrSymlink(src, dst, preserveInDst)
 }
@@ -301,10 +303,9 @@ type mountLookupFunc func(ps *LaidOutStructure) (string, error)
 // backup copies of files, newly created directories are removed
 type mountedFilesystemUpdater struct {
 	*MountedFilesystemWriter
-	backupDir         string
-	mountPoint        string
-	managedBootAssets []string
-	updateObserver    ContentObserver
+	backupDir      string
+	mountPoint     string
+	updateObserver ContentObserver
 }
 
 // newMountedFilesystemUpdater returns an updater for given filesystem
@@ -327,17 +328,11 @@ func newMountedFilesystemUpdater(rootDir string, ps *LaidOutStructure, backupDir
 	if err != nil {
 		return nil, fmt.Errorf("cannot find mount location of structure %v: %v", ps, err)
 	}
-	// find out if we need to preserve any boot assets
-	bootAssets, err := maybePreserveManagedBootAssets(mount, ps)
-	if err != nil {
-		return nil, fmt.Errorf("cannot preserve managed boot assets: %v", err)
-	}
 
 	fu := &mountedFilesystemUpdater{
 		MountedFilesystemWriter: fw,
 		backupDir:               backupDir,
 		mountPoint:              mount,
-		managedBootAssets:       bootAssets,
 		updateObserver:          observer,
 	}
 	return fu, nil
@@ -345,52 +340,6 @@ func newMountedFilesystemUpdater(rootDir string, ps *LaidOutStructure, backupDir
 
 func fsStructBackupPath(backupDir string, ps *LaidOutStructure) string {
 	return filepath.Join(backupDir, fmt.Sprintf("struct-%v", ps.Index))
-}
-
-func maybePreserveManagedBootAssets(mountPoint string, ps *LaidOutStructure) ([]string, error) {
-	if ps.Role != SystemSeed && ps.Role != SystemBoot {
-		return nil, nil
-	}
-	// TODO:UC20: this code should no longer be needed when we use
-	// the updater interface from boot, in particular we shouldn't
-	// try to find a bootloader from this place, we don't have
-	// quite enough info not to guess
-
-	// the assets are within the system-boot or system-seed partition, set
-	// the right flags so that files are looked for using their paths inside
-	// the partition
-	role := bootloader.RoleRecovery
-	if ps.Role == SystemBoot {
-		role = bootloader.RoleRunMode
-	}
-	opts := &bootloader.Options{
-		Role:        role,
-		NoSlashBoot: true,
-	}
-	bl, err := bootloader.Find(mountPoint, opts)
-	if err != nil {
-		if err == bootloader.ErrBootloader {
-			// no bootloader in the partition?
-			return nil, nil
-		}
-		return nil, err
-	}
-
-	tbl, ok := bl.(bootloader.TrustedAssetsBootloader)
-	if !ok {
-		// bootloader implementation does not support managing its
-		// assets
-		return nil, nil
-	}
-	managed, err := tbl.IsCurrentlyManaged()
-	if err != nil {
-		return nil, err
-	}
-	if !managed {
-		// assets are not managed
-		return nil, nil
-	}
-	return tbl.ManagedAssets(), nil
 }
 
 // entryDestPaths resolves destination and backup paths for given
@@ -426,8 +375,7 @@ func (f *mountedFilesystemUpdater) entrySourcePath(source string) string {
 // Update applies an update to a mounted filesystem. The caller must have
 // executed a Backup() before, to prepare a data set for rollback purpose.
 func (f *mountedFilesystemUpdater) Update() error {
-	preserveInDst, err := mapPreserve(f.mountPoint,
-		append(f.ps.Update.Preserve, f.managedBootAssets...))
+	preserveInDst, err := mapPreserve(f.mountPoint, f.ps.Update.Preserve)
 	if err != nil {
 		return fmt.Errorf("cannot map preserve entries for mount location %q: %v", f.mountPoint, err)
 	}
@@ -532,18 +480,22 @@ func (f *mountedFilesystemUpdater) updateOrSkipFile(dstRoot, source, target stri
 	}
 
 	if osutil.FileExists(dstPath) {
-		if strutil.SortedListContains(preserveInDst, dstPath) {
+		backupName := backupPath + ".backup"
+		sameStamp := backupPath + ".same"
+		preserveStamp := backupPath + ".preserve"
+
+		if strutil.SortedListContains(preserveInDst, dstPath) || osutil.FileExists(preserveStamp) {
 			// file is to be preserved
 			return ErrNoUpdate
 		}
-		if osutil.FileExists(backupPath + ".same") {
+		if osutil.FileExists(sameStamp) {
 			// file is the same as current copy
 			return ErrNoUpdate
 		}
-		if !osutil.FileExists(backupPath + ".backup") {
+		if !osutil.FileExists(backupName) {
 			// not preserved & different than the update, error out
 			// as there is no backup
-			return fmt.Errorf("missing backup file %q for %v", backupPath+".backup", target)
+			return fmt.Errorf("missing backup file %q for %v", backupName, target)
 		}
 	}
 
@@ -601,8 +553,7 @@ func (f *mountedFilesystemUpdater) Backup() error {
 		return fmt.Errorf("cannot create backup directory: %v", err)
 	}
 
-	preserveInDst, err := mapPreserve(f.mountPoint,
-		append(f.ps.Update.Preserve, f.managedBootAssets...))
+	preserveInDst, err := mapPreserve(f.mountPoint, f.ps.Update.Preserve)
 	if err != nil {
 		return fmt.Errorf("cannot map preserve entries for mount location %q: %v", f.mountPoint, err)
 	}
@@ -677,15 +628,42 @@ func (f *mountedFilesystemUpdater) checkpointPrefix(dstRoot, target string, back
 	return nil
 }
 
+func (f *mountedFilesystemUpdater) preserveChangeContent(change *ContentChange, backupPath string) error {
+	if change.Before == "" {
+		// a new file, nothing to preserve
+		return nil
+	}
+
+	preserveStamp := backupPath + ".preserve"
+	backupName := backupPath + ".backup"
+	sameStamp := backupPath + ".same"
+	if err := makeStamp(preserveStamp); err != nil {
+		return fmt.Errorf("cannot create a checkpoint file: %v", err)
+	}
+	for _, name := range []string{backupName, sameStamp} {
+		if err := os.Remove(name); err != nil && !os.IsNotExist(err) {
+			return fmt.Errorf("cannot remove existing stamp file: %v", err)
+		}
+	}
+	return nil
+}
+
 func (f *mountedFilesystemUpdater) observedBackupOrCheckpointFile(dstRoot, source, target string, preserveInDst []string, backupDir string) error {
 	change, err := f.backupOrCheckpointFile(dstRoot, source, target, preserveInDst, backupDir)
 	if err != nil {
 		return err
 	}
 	if change != nil {
-		dstPath, _ := f.entryDestPaths(dstRoot, source, target, backupDir)
-		if err := observe(f.updateObserver, ContentUpdate, f.ps, f.mountPoint, dstPath, change); err != nil {
+		dstPath, backupPath := f.entryDestPaths(dstRoot, source, target, backupDir)
+		act, err := observe(f.updateObserver, ContentUpdate, f.ps, f.mountPoint, dstPath, change)
+		if err != nil {
 			return fmt.Errorf("cannot observe pending file write %v\n", err)
+		}
+		if act == ChangePreserveBefore {
+			// observer asked for the change to be preserved
+			if err := f.preserveChangeContent(change, backupPath); err != nil {
+				return fmt.Errorf("cannot preserve change content: %v", err)
+			}
 		}
 	}
 	return nil
@@ -727,6 +705,7 @@ func (f *mountedFilesystemUpdater) backupOrCheckpointFile(dstRoot, source, targe
 		// the udpate, no need for backup
 		return changeNewFile, nil
 	}
+	// destination file exists beyond this point
 
 	if osutil.FileExists(backupName) {
 		// file already checked and backed up
@@ -740,11 +719,6 @@ func (f *mountedFilesystemUpdater) backupOrCheckpointFile(dstRoot, source, targe
 	// executed update pass
 
 	if strutil.SortedListContains(preserveInDst, dstPath) {
-		// file is to be preserved, create a relevant stamp
-		if !osutil.FileExists(dstPath) {
-			// preserve, but does not exist, will be written anyway
-			return changeNewFile, nil
-		}
 		if osutil.FileExists(preserveStamp) {
 			// already stamped
 			return nil, nil
@@ -835,8 +809,7 @@ func (f *mountedFilesystemUpdater) backupVolumeContent(volumeRoot string, conten
 func (f *mountedFilesystemUpdater) Rollback() error {
 	backupRoot := fsStructBackupPath(f.backupDir, f.ps)
 
-	preserveInDst, err := mapPreserve(f.mountPoint,
-		append(f.ps.Update.Preserve, f.managedBootAssets...))
+	preserveInDst, err := mapPreserve(f.mountPoint, f.ps.Update.Preserve)
 	if err != nil {
 		return fmt.Errorf("cannot map preserve entries for mount location %q: %v", f.mountPoint, err)
 	}
@@ -902,7 +875,8 @@ func (f *mountedFilesystemUpdater) rollbackFile(dstRoot, source, target string, 
 	preserveStamp := backupPath + ".preserve"
 
 	if strutil.SortedListContains(preserveInDst, dstPath) && osutil.FileExists(preserveStamp) {
-		// file was preserved at original location, do nothing
+		// file was preserved at original location by being
+		// explicitly listed
 		return nil
 	}
 	if osutil.FileExists(sameStamp) {
@@ -922,17 +896,24 @@ func (f *mountedFilesystemUpdater) rollbackFile(dstRoot, source, target string, 
 			return err
 		}
 	} else {
-		// none of the markers exists, file is not preserved, meaning, it has
-		// been added by the update
-		if err := os.Remove(dstPath); err != nil && !os.IsNotExist(err) {
-			return fmt.Errorf("cannot remove written update: %v", err)
+		// no backup marker, could be a new file or one that got
+		// preserved by request of the observer, in which case there
+		// will be a stamp
+		if !osutil.FileExists(preserveStamp) {
+			if err := os.Remove(dstPath); err != nil && !os.IsNotExist(err) {
+				return fmt.Errorf("cannot remove written update: %v", err)
+			}
+			// since it's a new file, there was no original content
+			data.Before = ""
+		} else {
+			// content was preserved at original location
+			data.Before = dstPath
 		}
-		// since it's a new file, there was no original content
-		data.Before = ""
 	}
 	// avoid passing source path during rollback, the file has been restored
 	// to the disk already
-	if err := observe(f.updateObserver, ContentRollback, f.ps, f.mountPoint, dstPath, data); err != nil {
+	_, err := observe(f.updateObserver, ContentRollback, f.ps, f.mountPoint, dstPath, data)
+	if err != nil {
 		return fmt.Errorf("cannot observe pending file rollback %v\n", err)
 	}
